Overhaul knockback, more parry work.

+++ src/main/java/me/anticode/abco/api/ExpandedAttack.java
@@ -4,8 +4,8 @@ public interface ExpandedAttack {
     void antisBetterCombatOverhauls$setCritical(boolean value);
     boolean antisBetterCombatOverhauls$getCritical();
 
-    void antisBetterCombatOverhauls$setKnockback(int value);
-    int antisBetterCombatOverhauls$getKnockback();
+    void antisBetterCombatOverhauls$setKnockback(double value);
+    double antisBetterCombatOverhauls$getKnockback();
 
     void antisBetterCombatOverhauls$setAttackSpeedMultiplier(double speed);
     double antisBetterCombatOverhauls$getAttackSpeedMultiplier();
